inventory bug fix, repair item proc added, changes on chaosbox

+++ WebZen.DataBase/SkillKeyDto.cs
@@ -26,6 +26,6 @@ namespace MU.DataBase
 
         public byte RkeyDefine { get; set; }
 
-        public uint QWERLevelDefine { get; set; }
+        public Int64 QWERLevelDefine { get; set; }
     }
 }
